requirements.txt
fastapi
uvicorn[standard]
motor
redis
faststream[rabbit]
pydantic
python-dotenv
